+++ log/query.go
@@ -23,7 +23,7 @@ func internalQuery(
 	) func(
 		trace.QueryDoDoneInfo,
 	) {
-		if d.Details()&trace.QueryPoolEvents == 0 {
+		if d.Details()&trace.QueryEvents == 0 {
 			return nil
 		}
 		ctx := with(*info.Context, TRACE, "ydb", "query", "do")
@@ -58,7 +58,7 @@ func internalQuery(
 					if !xerrors.IsYdb(info.Error) {
 						lvl = DEBUG
 					}
-					l.Log(WithLevel(ctx, lvl), "done",
+					l.Log(WithLevel(ctx, lvl), "failed",
 						latencyField(start),
 						Error(info.Error),
 						Int("attempts", info.Attempts),
@@ -75,7 +75,7 @@ func internalQuery(
 	) func(
 		trace.QueryDoTxDoneInfo,
 	) {
-		if d.Details()&trace.TablePoolAPIEvents == 0 {
+		if d.Details()&trace.QueryEvents == 0 {
 			return nil
 		}
 		ctx := with(*info.Context, TRACE, "ydb", "query", "do", "tx")
@@ -110,13 +110,70 @@ func internalQuery(
 					if !xerrors.IsYdb(info.Error) {
 						lvl = DEBUG
 					}
-					l.Log(WithLevel(ctx, lvl), "done",
+					l.Log(WithLevel(ctx, lvl), "failed",
 						latencyField(start),
 						Error(info.Error),
 						Int("attempts", info.Attempts),
 						versionField(),
 					)
 				}
+			}
+		}
+	}
+	t.OnCreateSession = func(info trace.QueryCreateSessionStartInfo) func(info trace.QueryCreateSessionDoneInfo) {
+		if d.Details()&trace.QuerySessionEvents == 0 {
+			return nil
+		}
+		ctx := with(*info.Context, TRACE, "ydb", "query", "session", "create")
+		l.Log(ctx, "start")
+		start := time.Now()
+
+		return func(info trace.QueryCreateSessionDoneInfo) {
+			if info.Error == nil {
+				l.Log(ctx, "done",
+					latencyField(start),
+					String("session_id", info.Session.ID()),
+					String("session_status", info.Session.Status()),
+				)
+			} else {
+				lvl := WARN
+				if !xerrors.IsYdb(info.Error) {
+					lvl = DEBUG
+				}
+				l.Log(WithLevel(ctx, lvl), "done",
+					latencyField(start),
+					Error(info.Error),
+					versionField(),
+				)
+			}
+		}
+	}
+	t.OnDeleteSession = func(info trace.QueryDeleteSessionStartInfo) func(info trace.QueryDeleteSessionDoneInfo) {
+		if d.Details()&trace.QuerySessionEvents == 0 {
+			return nil
+		}
+		ctx := with(*info.Context, TRACE, "ydb", "query", "session", "delete")
+		l.Log(ctx, "start",
+			String("session_id", info.Session.ID()),
+			String("session_status", info.Session.Status()),
+		)
+		start := time.Now()
+
+		return func(info trace.QueryDeleteSessionDoneInfo) {
+			if info.Error == nil {
+				l.Log(ctx, "done",
+					latencyField(start),
+				)
+			} else {
+				lvl := WARN
+				if !xerrors.IsYdb(info.Error) {
+					lvl = DEBUG
+				}
+				l.Log(WithLevel(ctx, lvl), "failed",
+					latencyField(start),
+					Error(info.Error),
+					versionField(),
+				)
 			}
 		}
 	}