+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+* Added metrics over query service internals
+* Added session create and delete events into `trace.Query`
+* Moved public type `query.SessionStatus` into `internal/query` package
+
 ## v3.57.2
 * Fixed cases when some option is nil